+++ docs/setup/changelog.rst
@@ -20,6 +20,7 @@ Upcoming Changes (main branch)
 - (Now really) added the ``logical_xor`` function.
 - Add ``nd.median``. An axis version is not yet implemented.
 - Added new namespaces :ref:`ndb <class_ndb>`, :ref:`ndf <class_ndf>` and :ref:`ndi <class_ndi>`, for full tensor reductions to ``bool``, ``float`` and ``int``, respectively.
+- Added ``NDArray.layout()``, through which you can inspect if an array has consistent strides.
 
 **Changed**